+++ manul/src/dev/session_parameters.rs
@@ -62,7 +62,7 @@ impl<D: digest::Digest> signature::DigestVerifier<D, TestSignature> for TestVeri
 #[derive(Debug, Clone, Copy, Default)]
 pub struct TestHasher {
     cursor: usize,
-    buffer: [u8; 32],
+    buffer: digest::Output<Self>,
 }
 
 impl digest::HashMarker for TestHasher {}
@@ -72,7 +72,7 @@ impl digest::Update for TestHasher {
         // A very simple algorithm for testing, just xor the data in buffer-sized chunks.
         for byte in data {
             *self.buffer.get_mut(self.cursor).expect("index within bounds") ^= byte;
-            self.cursor = (self.cursor + 1) % 32;
+            self.cursor = (self.cursor + 1) % self.buffer.len();
         }
     }
 }
@@ -84,7 +84,7 @@ impl digest::FixedOutput for TestHasher {
 }
 
 impl digest::OutputSizeUser for TestHasher {
-    type OutputSize = typenum::U8;
+    type OutputSize = typenum::U32;
 }
 
 /// An implementation of [`SessionParameters`] using the testing signer/verifier types.