+++ manul/src/protocol.rs
@@ -30,4 +30,5 @@ pub use round::{
 pub use serialization::{Deserializer, Serializer};
 
 pub(crate) use errors::ReceiveErrorType;
+pub(crate) use message::ProtocolMessagePartHashable;
 pub(crate) use object_safe::{BoxedRng, ObjectSafeRound};